pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-agents-evals"
version = "0.1.0"
description = "AI Agents Evaluation Tools"
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
    "openai",
    "pandas>=1.3.0",  # Ensure colalign parameter support
    "numpy",
    "pyyaml",
    "scipy",
    "tabulate",
    "azure-ai-projects>=2.0.0b1",
    "azure-identity",
    "aiohttp>=3.8.5",
    "packaging>=25.0"
]

[project.optional-dependencies]
dev = [
    "python-dotenv",
    "pytest",
    "pytest-snapshot",
    "pre-commit"
]

[tool.setuptools.packages.find]
where = ["analysis"]

[tool.isort]
profile = "black"
known_first_party = ["analysis", "tests"]
